==> M2 - Movie Profit Analysis/1. Rotten Tomatoes web scrape.ipynb <==
import os
import time
import pandas as pd
import requests
import numpy as np

from selenium import webdriver
from bs4 import BeautifulSoup

import warnings
warnings.filterwarnings('ignore')

chromedriver = "/Applications/chromedriver"
os.environ['webdriver.chrome.driver'] = chromedriver
driver = webdriver.Chrome(chromedriver)

# Open webpage for movies with a Tomatometer score of 70% or below
driver.get('https://www.rottentomatoes.com/browse/dvd-streaming-all?minTomato=0&maxTomato=70&services=amazon;hbo_go;itunes;netflix_iw;vudu;amazon_prime;fandango_now&genres=1;2;4;5;6;8;9;10;11;13;18;14&sortBy=release')

# Click "Show More" button
button = driver.find_element_by_xpath('/html/body/div[6]/div[2]/div[1]/div/div[2]/div[2]/div[3]/button')
button.click()

# Click "Show More" until page bottom is reached
last_height = driver.execute_script('return document.body.scrollHeight')
while True:
    button = driver.find_element_by_xpath('/html/body/div[6]/div[2]/div[1]/div/div[2]/div[2]/div[3]/button')
    button.click()
    time.sleep(4)
    new_height = driver.execute_script('return document.body.scrollHeight')
    if new_height == last_height:
        break
    last_height = new_height

# Import the HTML of the webpage into python  
soup = BeautifulSoup(driver.page_source, 'lxml')

# Grab the HTML of each movie
movie_card = soup.find_all('div', class_ = 'mb-movie') 

# Create a df to store scraped URLs and Tomatometer scores
url_tomatometer = pd.DataFrame({'Tomatoes URL':[''], 'Tomatometer':['']})

# Grab the URL for every movie on the page
for movie in movie_card:
    try:
        link = movie.find('a').get('href')
        tomatometer = movie.find('span', class_ = 'tMeterScore').text
        url_tomatometer = url_tomatometer.append({'Tomatoes URL':link, 'Tomatometer':tomatometer},ignore_index = True)
    except:
        pass
    
# Make full Rotten Tomatoes link in Link column
url_tomatometer['Tomatoes URL'] = 'https://www.rottentomatoes.com' + url_tomatometer['Tomatoes URL'].astype(str)

# Drop empty first row and reset index
url_tomatometer.drop(index = 0, inplace = True)
url_tomatometer = url_tomatometer.reset_index(drop=True)

# # Read in list of movies with low Tomatometer score
# url_tomatometer = pd.read_csv ('rottentomatoes70.csv',index_col=[0])

# url_tomatometer.info()
# url_tomatometer.tail(50)

# Add new columns to be filled in next Rotten Tomatoes scrape
url_tomatometer['Title'] = None
url_tomatometer['Year'] = None
url_tomatometer['N_Tomatometer'] = None

for index, row in url_tomatometer.iterrows():
    url = url_tomatometer.iloc[index,0] # Retrieves the IMDB url to search for
    page = requests.get(url)
    soup = BeautifulSoup(page.text, 'lxml') # Imports the HTML of the webpage into python 
    time.sleep(2) 
            
    scoreboard = soup.find('score-board',class_='scoreboard')

    try:
        title=scoreboard.find('h1',{'data-qa':'score-panel-movie-title'}).text  
        url_tomatometer['Title'][index]=title
    except:
        title = np.nan 
    
    try:
        year=scoreboard.find('p',class_='scoreboard__info').text 
        url_tomatometer['Year'][index]=year
    except:
        year = np.nan
    
    try:
        n_tomatometer=scoreboard.find('a',{'data-qa':'tomatometer-review-count'}).text 
        url_tomatometer['N_Tomatometer'][index]=n_tomatometer
    except:
        n_tomatometer = np.nan
    
# Export the dataframe as a csv
url_tomatometer.to_csv('rotten_df.csv')
