==> M2 - Movie Profit Analysis/2. Rotten Tomatoes - IMDB dataset merge.ipynb <==
import os
import pandas as pd
import requests
import numpy as np
import time

from bs4 import BeautifulSoup
from selenium import webdriver
from selenium.webdriver.common.by import By
from selenium.webdriver.common.keys import Keys

# from selenium.webdriver.support.ui import WebDriverWait
# from selenium.webdriver.support import expected_conditions as EC


# Read in list of movies with low Tomatometer score
rotten_df = pd.read_csv ('rotten_df.csv',index_col=[0])

# Display full column width (for the links)
pd.set_option('display.max_colwidth', None)

# Remove non-numeric info from end of Year and N_Tomatometer columns
rotten_df['Year'] = rotten_df['Year'].str[:4]
rotten_df['N_Tomatometer'] = rotten_df['N_Tomatometer'].str[:-8]
rotten_df

rotten_df.info()

# Datasets found on IMDB site contain data https://www.imdb.com/interfaces/

# Read in tsv file to match movie titles to IMDB unique ID (tconst)
basics_tsv = pd.read_csv('title_basics.tsv', delimiter="\t")

# Data on IMDB ratings / number of votes
ratings_tsv = pd.read_csv('title_ratings.tsv', delimiter="\t")

# # Data on movie directors / writers
# crew_tsv = pd.read_csv('title_crew.tsv', delimiter="\t")

# # Data on movie cast / crew
# principals_tsv = pd.read_csv('title_principals.tsv', delimiter="\t")

basics_tsv['titleType'].value_counts()

## Clean up basics_tsv file

# Select for only movies and videos
basics_tsv = basics_tsv[(basics_tsv['titleType']=='movie') | (basics_tsv['titleType']=='video')]
# Drop unnecessary columns
IMDB_ID_df = basics_tsv.drop(columns=['titleType','originalTitle', 'endYear'])
# Rename title and year column
IMDB_ID_df.rename(columns={"primaryTitle": "Title", "startYear": "Year"},inplace=True)

# Convert year to int
# IMDB_ID_df['startYear']=pd.to_numeric(IMDB_ID_df['startYear'], errors='coerce')
# IMDB_ID_df['startYear'] = IMDB_ID_df['startYear'].astype(str).astype(int,errors='ignore')

IMDB_ID_df.info()

ratings_tsv.sample(10)

# Merge Rotten Tomatoes df with IMDB_ID df on two columns: Year and Title
merged_df = rotten_df.merge(IMDB_ID_df, on=["Year", "Title"])

merged_df.info()

# Merge Rotten Tomatoes df with ratings_tsv df on tconst column
merged_df = merged_df.merge(ratings_tsv, on=["tconst"])

merged_df.info()

merged_df.head(50)

merged_df.to_csv('merged_df.csv')

### Search and scrape Box Office Mojo for gross revenue values

merged_df=pd.read_csv('merged_df.csv')

# Create new column with full Box Office Mojo url
merged_df['BOM url'] = "https://www.boxofficemojo.com/title/" + merged_df['tconst']+"/"
merged_df

# Prep for incoming IMDB data
merged_df['domestic'] = None
merged_df['international'] = None
merged_df['worldwide'] = None
merged_df['domestic_distributor'] = None
merged_df['domestic_opening'] = None
merged_df['mpaa'] = None
merged_df['Genres'] = None

merged_df

short_df = merged_df[7000:]
short_df.reset_index(inplace=True,drop=True)
for index, row in short_df.iterrows():
    url = short_df.loc[index,'BOM url'] # Retrieves the IMDB url to search for
    page = requests.get(url)
    soup = BeautifulSoup(page.text, 'lxml') # Imports the HTML of the webpage into python 
    time.sleep(3) 
    
    revenue_box = soup.find('div',class_='a-section a-spacing-none mojo-performance-summary-table')
    
    try:
        revenue_box.find_all('span',class_=['money','percent zero'])
        domestic=revenue_box.find_all('span',class_=['money','percent zero'])[0].text      
        international=revenue_box.find_all('span',class_=['money','percent zero'])[1].text
        worldwide=revenue_box.find_all('span',class_=['money','percent zero'])[2].text
    except:
        domestic = np.nan
        international = np.nan
        worldwide = np.nan
    short_df['domestic'][index]=domestic
    short_df['international'][index]=international
    short_df['worldwide'][index]=worldwide
    try:
        domestic_distributor = soup.find(string = 'Domestic Distributor').findNext().text
    except:
        domestic_distributor = np.nan
    short_df['domestic_distributor'][index]=domestic_distributor
    try:
        domestic_opening = soup.find(string = 'Domestic Opening').findNext().text
    except:
        domestic_opening = np.nan
    short_df['domestic_opening'][index]=domestic_opening
    try:
        mpaa = soup.find(string = 'MPAA').findNext().text
    except:
        mpaa = np.nan
    short_df['mpaa'][index]=mpaa
    try:
        genres = soup.find(string = 'Genres').findNext().text
    except:
        genres = np.nan
    short_df['Genres'][index]=genres


short_df.to_csv('full_7000_7990.csv')

short_df.info()

short_df = merged_df[3000:4000]
short_df.reset_index(inplace=True,drop=True)
short_df.iloc[633]
