==> M2 - Movie Profit Analysis/3. Dataframe merges and data cleaning.ipynb <==
import pandas as pd
import numpy as np


# Display full column width (for the links)
pd.set_option('display.max_colwidth', None)

def read_csvs(filelist):
    df = pd.DataFrame()
    for file in filelist:
        sub_df = pd.read_csv(file, index_col=[0])
        df = pd.concat([df,sub_df])
    return df

filelist = ['full_0_1999.csv', 'full_2000_2999.csv', 'full_3000_4999.csv', 'full_5000_6999.csv', 'full_7000_7990.csv']
df = read_csvs(filelist)
df

# drop a couple of columns that are not useful
df.drop(['Unnamed: 0.1'],axis=1,inplace = True)
df.drop(['genres'], axis = 1, inplace = True)

## Clean final movies dataframe

columns_to_clean = ['Tomatometer', 'Year', 'N_Tomatometer', 'isAdult', 'runtimeMinutes', 
                    'domestic', 'international', 'worldwide', 'domestic_opening', 
                    'averageRating', 'numVotes']

for column in columns_to_clean:
    if column == 'Tomatometer':
        df[column] = df[column].str[:-1]
    if column == 'runtimeMinutes':
        df[column].replace(to_replace=r"\N", value='', inplace=True)
    df=df.replace('–', np.nan)
    df[column] = df[column].astype(str)
    df[column] = df[column].str.replace('%', '')
    df[column] = df[column].str.replace('$', '')
    df[column] = df[column].str.replace(',', '')
    df[column] = pd.to_numeric(df[column], errors='coerce') # coerce to ignore NaN values

# Rename columns
df=df.rename(columns={"domestic": "Domestic_Gross", 
                   "international": "International_Gross", 
                   "worldwide": "Worldwide_Gross",
                   "domestic_distributor": "Domestic_Distributor",
                   "domestic_opening": "Domestic_Opening",
                   "mpaa": "MPAA_Rating",
                   "averageRating": "Average_Audience_Rating",
                   "numVotes": "Number_Audience_Votes"})

df.info()


df.head()

# save to pickle
df.to_pickle("prep_data.pkl")

# Create additional boolean columns for movie genres
df['Genres']=df['Genres'].str.replace('\n    \n        ', ',') # clean up newlines / spaces and replace with commas
dummy_genres = df['Genres'].str.get_dummies(',')
df = pd.concat([df, dummy_genres], axis=1)
df.shape

# Fill all unrated movies with string 'Unrated'
df['MPAA_Rating'] = df['MPAA_Rating'].fillna('Unrated')

# Create additional boolean columns for MPAA ratings (or unrated)
MPAA = df['MPAA_Rating'].str.get_dummies(' ')
df = pd.concat([df, MPAA], axis=1)
# df = df.join(MPAA)
df.shape

df.to_csv('data_for_analysis.csv')
